LayoutTests/platform/glib/imported/w3c/web-platform-tests/reporting/same-origin-same-site-credentials.https.sub-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'report-to'.

CONSOLE MESSAGE: Refused to load https://web-platform.test:9443/reporting/resources/fail.png because it does not appear in the img-src directive of the Content Security Policy.


FAIL Reporting endpoints received credentials. assert_unreached: A report of csp-violation from https://web-platform.test:9443/reporting/resources/same-origin-report.https.sub.html is not found. Reached unreachable code
